Changed QWizardPages into seperate child classes to reduce confusion and bugginess when running
