--- QtObjects/WrappedLabel.py
@@ -1,6 +1,0 @@
-from PyQt6.QtWidgets import QLabel
-
-class WrappedLabel(QLabel):
-    def __init__(self, text):
-        super().__init__(text)
-        self.setWordWrap(True)
