+++ main.py
@@ -1,20 +1,14 @@
 import sys
 from PyQt6.QtWidgets import QApplication, QWizard
 
-from QtObjects.WizardPage import WizardPage
-from subpageContents import (
-    introPage, importSpreadsheet, selectCourses,
-    customizeDetails, reviewInfo, generateSchedule
-)
+from QtObjects.WizardPageChildren import Introduction, ImportSchedule, SelectCourses, CustomizeDetails, ReviewInfo, GenerateSchedule
 
-# KEY: Index (int)
-# VALUE: ['Page Title', Page Create Function, Reset Page Layout When Backtracking]
-WIZARD_PAGES = {0: ['Introduction', introPage, False],
-                1: ['Import Spreadsheet', importSpreadsheet, False],
-                2: ['Select Courses', selectCourses, True],
-                3: ['Customize Details', customizeDetails, False],
-                4: ['Review Information', reviewInfo, True],
-                5: ['Generate Schedule', generateSchedule, False]}
+WIZARD_PAGES = [['Introduction', Introduction],
+                ['Import Spreadsheet', ImportSchedule],
+                ['Select Courses', SelectCourses],
+                ['Customize Details', CustomizeDetails],
+                ['Review Information', ReviewInfo],
+                ['Generate Schedule', GenerateSchedule]]
 
 if __name__ == '__main__':
     app = QApplication(sys.argv)
@@ -23,15 +17,14 @@ if __name__ == '__main__':
     wizard = QWizard()
     wizard.setWindowTitle('Simmons University Finals Exam Schedule Wizard')
     wizard.setWizardStyle(QWizard.WizardStyle.ModernStyle)
-    wizard.resize(app.primaryScreen().size() / 2.5)
+    wizard.resize(app.primaryScreen().size() / 2.25)
 
     # Make list of all page titles
-    pageTitles = [WIZARD_PAGES[key][0] for key in sorted(WIZARD_PAGES)]
+    pageTitles = [page[0] for page in WIZARD_PAGES]
 
-    for key in sorted(WIZARD_PAGES.keys()):
-        func = WIZARD_PAGES[key][1]
-        resetPage = WIZARD_PAGES[key][2]
-        wizard.addPage(WizardPage(pageTitles, key, func, resetPage))
+    for index in range(len(WIZARD_PAGES)):
+        wizardPage = WIZARD_PAGES[index][1]
+        wizard.addPage(wizardPage(pageTitles, index))
 
     # Show the wizard and run event loop
     wizard.show()
